fix: address compilation issues for non-linux oses
