+++ auditbeat/module/file_integrity/eventreader_linux.go
@@ -1,0 +1,44 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+//go:build linux
+
+package file_integrity
+
+import (
+	"github.com/elastic/elastic-agent-libs/logp"
+)
+
+// NewEventReader creates a new EventProducer backed by fsnotify.
+func NewEventReader(c Config) (EventProducer, error) {
+	switch c.ForceBackend {
+	case BackendKProbes:
+		return &kProbesReader{
+			config:  c,
+			log:     logp.NewLogger(moduleName),
+			parsers: FileParsers(c),
+		}, nil
+	case BackendFSNotify:
+		fallthrough
+	default:
+		return &fsNotifyReader{
+			config:  c,
+			log:     logp.NewLogger(moduleName),
+			parsers: FileParsers(c),
+		}, nil
+	}
+}
